refact: abstract processInviteMembersInput function
